split Destination enum into unit
